texture_generation/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-limiter
flasgger
python-dotenv
psycopg2-binary
opencv-python
numpy
Pillow
roboflow
inference
diffusers
transformers
accelerate
safetensors

